+++ src/components/AdminMenu.php
@@ -165,37 +165,32 @@ class AdminMenu extends \yii\base\Component
             // check if this is an entrie with a permission
             if ($item['permissionIsRoute']) {
                 // verify if the permission is provided for this user:
-                // if the permission is granted will write inti $responseData,
+                // if the permission is granted will write in $responseData,
                 // if not we continue;
                 if (!Yii::$app->auth->matchRoute($this->getUserId(), $item['permissionRoute'])) {
                     continue;
                 }
             }
 
+            $authIds = [];
             // this item does have groups
             if (isset($item['groups'])) {
                 $permissionGranted = false;
 
                 // see if the groups has items
                 foreach ($item['groups'] as $groupName => $groupItem) {
-                    if (count($groupItem['items'])  > 0) {
-                        if ($permissionGranted) {
-                            continue;
-                        }
-
+                    if (count($groupItem['items']) > 0) {
                         foreach ($groupItem['items'] as $groupItemEntry) {
-                            // a previous entry already has solved the question if the permission is granted
-                            if ($permissionGranted) {
-                                continue;
-                            }
                             if ($groupItemEntry['permissionIsRoute']) {
                                 // when true, set permissionGranted to true
-                                if (Yii::$app->auth->matchRoute($this->getUserId(), $groupItemEntry['route'])) {
+                                if (($id = Yii::$app->auth->matchRoute($this->getUserId(), $groupItemEntry['route']))) {
+                                    $authIds[] = $id;
                                     $permissionGranted = true;
                                 }
                             } elseif ($groupItemEntry['permissionIsApi']) {
                                 // when true, set permissionGranted to true
-                                if (Yii::$app->auth->matchApi($this->getUserId(), $groupItemEntry['permissionApiEndpoint'])) {
+                                if (($id = Yii::$app->auth->matchApi($this->getUserId(), $groupItemEntry['permissionApiEndpoint']))) {
+                                    $authIds[] = $id;
                                     $permissionGranted = true;
                                 }
                             } else {
@@ -205,6 +200,7 @@ class AdminMenu extends \yii\base\Component
                     }
                 }
 
+                // skip menu stack for this item
                 if (!$permissionGranted) {
                     continue;
                 }
@@ -221,12 +217,12 @@ class AdminMenu extends \yii\base\Component
             $responseData[] = [
                 'moduleId' => $item['moduleId'],
                 'id' => $item['id'],
+                'authIds' => $authIds,
                 'template' => $item['template'],
                 'routing' => $item['routing'],
                 'alias' => $alias,
                 'icon' => $item['icon'],
                 'searchModelClass' => $item['searchModelClass'],
-
             ];
         }
 
@@ -263,12 +259,14 @@ class AdminMenu extends \yii\base\Component
                 }
                 foreach ($groupItem['items'] as $groupItemKey => $groupItemEntry) {
                     if ($groupItemEntry['permissionIsRoute']) {
-                        if (!Yii::$app->auth->matchRoute($this->getUserId(), $groupItemEntry['route'])) {
+                        $id = Yii::$app->auth->matchRoute($this->getUserId(), $groupItemEntry['route']);
+                        if (!$id) {
                             unset($data['groups'][$groupName]['items'][$groupItemKey]);
                             continue;
                         }
                     } elseif ($groupItemEntry['permissionIsApi']) {
-                        if (!Yii::$app->auth->matchApi($this->getUserId(), $groupItemEntry['permissionApiEndpoint'])) {
+                        $id = Yii::$app->auth->matchApi($this->getUserId(), $groupItemEntry['permissionApiEndpoint']);
+                        if (!$id) {
                             unset($data['groups'][$groupName]['items'][$groupItemKey]);
                             continue;
                         }
@@ -290,6 +288,7 @@ class AdminMenu extends \yii\base\Component
                     $data['groups'][$groupName]['items'][$groupItemKey]['hiddenInMenu'] = AdminMenuBuilder::getOptionValue($groupItemEntry, 'hiddenInMenu', false);
                     $data['groups'][$groupName]['items'][$groupItemKey]['pool'] = AdminMenuBuilder::getOptionValue($groupItemEntry, 'pool', null);
                     $data['groups'][$groupName]['items'][$groupItemKey]['alias'] = $alias;
+                    $data['groups'][$groupName]['items'][$groupItemKey]['authId'] = $id;
                 }
                 
                 // if there are no items for this group, unset the group from the data array